add tests for Bucket.remove
