+++ tests/bucket_test.py
@@ -1,0 +1,22 @@
+"""Tests for Bucket"""
+import pytest
+
+from reduct import ReductError
+
+
+@pytest.mark.asyncio
+async def test__remove_ok(client):
+    """Should remove a bucket"""
+    bucket = await client.create_bucket("bucket")
+    await bucket.remove()
+    with pytest.raises(ReductError):
+        await client.get_bucket("bucket")
+
+
+@pytest.mark.asyncio
+async def test__remove_not_exist(client):
+    """Should not remove a bucket if it doesn't exist"""
+    bucket = await client.create_bucket("bucket")
+    await bucket.remove()
+    with pytest.raises(ReductError):
+        await bucket.remove()
