+++ tests/client_test.py
@@ -1,43 +1,11 @@
-"""Just stub for pipline"""
-import random
+"""Tests for Client"""
 from asyncio import sleep
 from typing import List
 
 import pytest
 from aiohttp import ClientConnectionError
 
-from reduct import Client, ServerInfo, BucketList, ReductError, BucketInfo, Bucket
-
-
-@pytest.fixture(name="url")
-def _url() -> str:
-    return "http://127.0.0.1:8383"
-
-
-@pytest.fixture(name="client")
-def _make_client(url):
-    client = Client(url)
-    return client
-
-
-@pytest.fixture(name="bucket_1")
-async def _bucket_1(client) -> Bucket:
-    bucket = await client.create_bucket("bucket-1")
-    await bucket.write("entry-1", b"some-data-1", timestamp=1)
-    await bucket.write("entry-1", b"some-data-2", timestamp=2)
-    await bucket.write("entry-2", b"some-data-3", timestamp=3)
-    await bucket.write("entry-2", b"some-data-4", timestamp=4)
-    yield bucket
-    await bucket.remove()
-
-
-@pytest.fixture(name="bucket_2")
-async def _bucket_2(client) -> Bucket:
-    bucket = await client.create_bucket("bucket-2")
-    await bucket.write("entry-1", b"some-data-1", timestamp=5)
-    await bucket.write("entry-1", b"some-data-2", timestamp=6)
-    yield bucket
-    await bucket.remove()
+from reduct import Client, ServerInfo, ReductError, BucketInfo
 
 
 @pytest.mark.asyncio
